Added python installation to cmake scripts
This commit added a CMakeLists.txt to the lib/python which pretty much
just tells cmake to call 'python ./setup.py install prefix=cmake_dir'
when installing.

Also added a line to lib/CMakeLists.txt to include lib/python in
the cmake project.

A known bug bug introduced in this commit is that cmake generates
a build folder in lib/python. This is presumably for storing temporary
files during the compilation of the python module. Consider altering
some of the options in the EXECUTE_PROCESS part of python's CMakeLists
to either destroy the build directory after the build is done or add
a directive to clean up the build directory during a make clean.