Added a cleaning step to the python cmake script
This commit adds a few lines to lib/python/CMakeLists.txt
that tell cmake to clean up the python source directory after it
finishes installing the python module.

It IS a littly hacky and could be better done by copying the
whole lib/python folder to cmake's staging area that would get
cleaned up automatically, but this works fine.